requirements.txt
flask
kafka-python
scapy